sales_rep_tableau_ETL/requirements.txt
asn1crypto==1.5.1
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
filelock==3.15.4
idna==3.7
kaggle==1.6.14
numpy==2.0.0
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
six==1.16.0
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomlkit==0.13.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
